feat: Add TasksContext
Created TasksContext and TasksProvider components to manage tasks state. The TasksContext provides a tasks array and a setTasks function through context, allowing other components to access and update the tasks data. The TasksProvider initializes the tasks state using the data stored in the localStorage if available. It also updates the localStorage whenever the tasks state changes to keep it in sync.