refactor: changed modal functionality
Updated the modal functionality in the component. Replaced the previous state variables and handlers with a new modalData state object that contains properties for different modal states (deleteModalOpen and editModalOpen).

Introduced the openModal function, which takes a modalName parameter and updates the modalData state by setting the corresponding modal property to true.

Also added the closeModal function, which takes a modalName parameter and updates the modalData state by setting the corresponding modal property to false.

This refactoring allows for more flexible and scalable management of modal states in the component.